backend/requirements.txt
djangorestframework
asgiref==3.10.0
Django==5.2.8
django-cors-headers==4.9.0
pillow==12.0.0
psycopg2-binary==2.9.11
python-dotenv==1.2.1
sqlparse==0.5.3